Added support for adding version to elements in collections.
Attempts to get the class type from the field name, which succeeds for
tag names such as <list>, but List.class does not have any fields, so
attempts to get the element type fail.

The name of the element members is based on the alias for that class,
which can be identified from the mapper.

More documentation to come.